feat(heuristics): probe for the biggest cube input given a device
