+++ plantseg/functionals/predictions/utils/array_predictor.py
@@ -11,172 +11,10 @@ from plantseg.functionals.predictions.utils.array_dataset import (
     default_prediction_collate,
     remove_padding,
 )
+from plantseg.functionals.predictions.utils.size_finder import _is_2d_model, find_batch_size, will_CUDA_OOM
 from plantseg.training.embeddings import embeddings_to_affinities
-from plantseg.training.model import UNet2D
 
 logger = logging.getLogger(__name__)
-
-
-def _is_2d_model(model: nn.Module) -> bool:
-    if isinstance(model, nn.DataParallel):
-        model = model.module
-    return isinstance(model, UNet2D)
-
-
-def find_patch_and_halo_shapes(
-    full_volume_shape: tuple[int, int, int],
-    max_patch_shape: tuple[int, int, int],
-    min_halo_shape: tuple[int, int, int],
-) -> tuple[tuple[int, int, int], tuple[int, int, int]]:
-    """
-    Recommend patch shape and halo size for a given 3D sample shape.
-
-    Args:
-        full_volume_shape (tuple[int, int, int]): Shape of the entire 3D sample.
-        max_patch_shape (tuple[int, int, int]): Maximum feasible patch shape for the selected GPU.
-        min_halo_shape (tuple[int, int, int]): Minimum halo size, counting both sides.
-
-    Returns:
-        tuple[tuple[int, int, int], tuple[int, int, int]]: Recommended patch shape and halo shape.
-    """
-    full_volume_shape = np.array(full_volume_shape)
-    max_patch_shape = np.array(max_patch_shape)
-    min_halo_shape = np.array(min_halo_shape)
-
-    n_voxels_patch = np.prod(max_patch_shape)
-    n_voxels_sample = np.prod(full_volume_shape)
-
-    if n_voxels_patch >= n_voxels_sample:
-        return tuple(full_volume_shape), (0, 0, 0)
-
-    # Adjust patch shape if necessary
-    shrink = max_patch_shape > full_volume_shape
-    adjusted_patch_shape = np.minimum(full_volume_shape, max_patch_shape)
-
-    if np.any(shrink):
-        n_shrink_dims = np.sum(shrink)
-        if n_shrink_dims == 2:
-            remaining_dim = np.flatnonzero(~shrink)
-            adjusted_patch_shape[remaining_dim] = min(
-                n_voxels_patch // np.prod(full_volume_shape[shrink]), full_volume_shape[remaining_dim]
-            )
-        elif n_shrink_dims == 1:
-            remaining_dims = np.flatnonzero(~shrink)
-            max_area = n_voxels_patch // full_volume_shape[shrink]
-            if max_area <= np.prod(full_volume_shape[remaining_dims]):
-                adjusted_patch_shape[remaining_dims] = np.sqrt(max_area).astype(int)
-            else:
-                return tuple(full_volume_shape), (0, 0, 0)
-    halo_shape = np.where(shrink, 0, min_halo_shape)
-    return tuple(adjusted_patch_shape - halo_shape), tuple(halo_shape)
-
-
-def find_batch_size(
-    model: nn.Module,
-    in_channels: int,
-    patch_shape: tuple[int, int, int],
-    patch_halo: tuple[int, int, int],
-    device: str,
-) -> int:
-    """Determine the maximum feasible batch size for a given model based on available GPU memory.
-
-    Args:
-        model (nn.Module): The model to be used for predictions.
-        in_channels (int): Number of input channels to the model.
-        patch_shape (tuple[int, int, int]): Dimensions of the input patches.
-        patch_halo (tuple[int, int, int]): Halo size used for patch augmentation.
-        device (str): Device to perform the computation on ('cuda' or 'cpu').
-
-    Returns:
-        int: The largest batch size that can be used without causing memory overflow.
-    """
-    if device == "cpu":
-        return 1
-
-    actual_patch_shape = tuple(patch_shape[i] + 2 * patch_halo[i] for i in range(3))
-    if isinstance(model, UNet2D):
-        actual_patch_shape = actual_patch_shape[1:]
-
-    model = model.to(device)
-    model.eval()
-    with torch.no_grad():
-        for batch_size in [1, 2, 4, 8, 16, 32, 64, 128]:
-            try:
-                x = torch.randn((batch_size, in_channels) + actual_patch_shape).to(device)
-                _ = model(x)
-            except RuntimeError as e:
-                if "out of memory" in str(e):
-                    batch_size //= 2
-                    break
-                else:
-                    raise
-            finally:
-                del x
-                torch.cuda.empty_cache()
-    if batch_size == 0:
-        raise RuntimeError(
-            f"Could not determine a feasible batch size for patch size {patch_shape} and halo {patch_halo}. "
-            "Please reduce the patch size."
-        )
-    del model
-    torch.cuda.empty_cache()
-    return batch_size
-
-
-def will_CUDA_OOM(
-    model: nn.Module,
-    in_channels: int,
-    patch_shape: tuple[int, int, int],
-    patch_halo: tuple[int, int, int],
-    batch_size: int,
-    device: str,
-) -> bool:
-    """Determine if a given batch size will cause an out-of-memory (OOM) error on the specified device.
-
-    Args:
-        model (nn.Module): The model to be used for predictions.
-        in_channels (int): Number of input channels to the model.
-        patch_shape (tuple[int, int, int]): Dimensions of the input patches.
-        patch_halo (tuple[int, int, int]): Halo size used for patch augmentation.
-        batch_size (int): Number of samples per batch.
-        device (str): Device to perform the computation on ('cuda' or 'cpu').
-
-    Returns:
-        bool: True if the batch size will cause an OOM error, False otherwise.
-    """
-    if device == "cpu":
-        return False  # CPU does not have CUDA OOM errors
-
-    # Calculate the actual patch shape including the halo
-    actual_patch_shape = tuple(patch_shape[i] + 2 * patch_halo[i] for i in range(3))
-
-    # If the model is a 2D UNet, adjust the patch shape to 2D
-    if isinstance(model, UNet2D):
-        actual_patch_shape = actual_patch_shape[1:]
-
-    model = model.to(device)
-    model.eval()
-
-    OOM_error = False
-    x = None
-    try:
-        with torch.no_grad():
-            x = torch.randn((batch_size, in_channels) + actual_patch_shape).to(device)
-            _ = model(x)
-    except RuntimeError as e:
-        if "out of memory" in str(e):
-            OOM_error = True
-            logger.info(
-                f"Using patch shape {patch_shape}, halo {patch_halo}, and batch size {batch_size} will cause OOM."
-            )
-        else:
-            raise  # Re-raise if it's not an OOM error
-    finally:
-        del x
-        del model
-        torch.cuda.empty_cache()
-
-    return OOM_error
 
 
 class ArrayPredictor:
